requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
tiktoken==0.5.2

# Vector Database and Embeddings
chromadb==0.4.18
pinecone-client==2.2.4
faiss-cpu==1.7.4

# LlamaIndex
llama-index==0.9.30
llama-index-vector-stores-chroma==0.1.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Streamlit Frontend
streamlit==1.29.0
streamlit-authenticator==0.2.3

# Web Search and External APIs
tavily-python==0.3.3
requests==2.31.0

# Monitoring and Observability
langsmith==0.0.69
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.0
pandas==2.1.4
numpy==1.24.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1